refactor: remove *_new suffix from load model files
