+++ geojson_modelica_translator/model_connectors/load_connectors/__init__.py
@@ -27,6 +27,6 @@ IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISI
 OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ****************************************************************************************************
 """
-from .spawn_new import Spawn  # noqa: F401
-from .teaser_new import Teaser  # noqa: F401
-from .time_series_new import TimeSeries  # noqa: F401
+from .spawn import Spawn  # noqa: F401
+from .teaser import Teaser  # noqa: F401
+from .time_series import TimeSeries  # noqa: F401
