+++ tests/model_connectors/test_district_system_new.py
@@ -48,7 +48,7 @@ from geojson_modelica_translator.model_connectors.energy_transfer_systems.coolin
 from geojson_modelica_translator.model_connectors.energy_transfer_systems.ets_hot_water_stub import (
     EtsHotWaterStub
 )
-from geojson_modelica_translator.model_connectors.load_connectors.time_series_new import (
+from geojson_modelica_translator.model_connectors.load_connectors.time_series import (
     TimeSeries
 )
 from geojson_modelica_translator.model_connectors.networks.network_chilled_water_stub import (
