+++ tests/model_connectors/test_time_series_heating_indirect.py
@@ -48,7 +48,7 @@ from geojson_modelica_translator.model_connectors.energy_transfer_systems.ets_co
 from geojson_modelica_translator.model_connectors.energy_transfer_systems.heating_indirect import (
     HeatingIndirect
 )
-from geojson_modelica_translator.model_connectors.load_connectors.time_series_new import (
+from geojson_modelica_translator.model_connectors.load_connectors.time_series import (
     TimeSeries
 )
 from geojson_modelica_translator.model_connectors.networks.network_heated_water_stub import (
